Split test_actions.py into separate actions
